Add files via upload
Base structure to create a hierarchy system of communications between the sensors. This primary file is used to recognise the DHT11 and manipulate his data using c++ pointers. A couple of lines in especial into extern "C" is write on classic C programming language, if you notice the line extern is used to tell the compiler that he need use a C Linker.